pyproject.toml
[project]
name = "multiagents"
version = "0.1.0"
description = "This is a lab to explore sql agents"
authors = [{ name = "Pilar Madariaga", email = "pilar.madariaga@gft.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
dependencies = [
    "agno>=1.1.0",
    "chromadb>=0.6.3",
    "composio-agno>=0.7.2",
    "duckdb>=1.2.0",
    "duckduckgo-search>=7.3.2",
    "google-search-results>=2.4.2",
    "gradio>=4.44.1",
    "groq>=0.18.0",
    "huggingface-hub>=0.28.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "lancedb>=0.19.0",
    "langchain-community>=0.3.17",
    "nbformat>=5.10.4",
    "ollama>=0.4.7",
    "openai>=1.62.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "pypdf>=5.3.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
]



[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",

    "mypy>=0.991",

    "ruff>=0.9.2",

]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]